Add rule to basic matching
This new rule removes all matched common settings from a configuration for which a
user has app-specific settings defined. Example user preferences:
* common: high contrast
* app-specific: font-size for Chrome

The core matching would also add high contrast to the inferred
configuration of Chrome. But this might be not the expected behavior of
the user. Hence, no further settings for Chrome are inferred.